// src/types.ts
export type Grade = 1 | 2 | 3 | 4 | 5;
// ...
export interface LessonActivity {
  name: string; // 1. Khởi động, 2. Khám phá, 3. Luyện tập, 4. Vận dụng
  objective: string; // Mục tiêu
  teacherActivity: string; // Hoạt động của giáo viên (cột 1)
  studentActivity: string; // Hoạt động của học sinh (cột 2)
}